feat(demo): parse the committed trace ZIP in-browser for the locator picker
The demo's dom-snapshot handler served a hand-pasted, pre-rendered copy of the
committed demo trace (stale-prone, and lacking the recorded viewport), so the
locator picker mostly fell back to the ARIA render. Plug it into the real
trace file instead: the service worker now fetches
public/demo/traces/checkout-pay-timeout.zip, inflates it with
DecompressionStream('deflate-raw') via a DataView central-directory walk
(app/demo/trace-zip.client.ts), and runs the exact same parsing/rendering the
server uses — so the picker and DOM snapshot card show the genuine
failure-time DOM, proportion-preserving zoom included, and re-recording the
trace needs no manual JSON pasting.

- Split the node-free pieces out of the trace pipeline so the browser can
  import them without pulling in zlib/storage: trace-events.ts (event model +
  JSONL parsing) and dom-snapshot-render.ts (snapshot rendering/sanitizing);
  trace-parser.ts and dom-snapshot.ts keep the node-only loading. No
  re-exports — Nitro auto-imports every server/utils export and duplicates
  shadow each other.
- Demo handler mirrors resolveCaseDomSnapshot: trace first (parsed once per
  path, failed loads uncached), then the seeded ARIA fallback.
- Unit tests: browser ZIP path renders the committed trace byte-for-byte
  identically to the server zlib path, and guards viewport presence so
  re-recorded traces keep working.

Verified in the built demo (service worker + sql.js): the trace-bearing
execution and cluster pages render "from the failure-time DOM snapshot" with
the 640x480 viewport and the real Pay button; trace-less clusters still get
the ARIA fallback.
